[Bronze V] Title: 이상한 기호, Time: 144 ms, Memory: 15904 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -13,8 +13,8 @@ public class Main{
         
         st = new StringTokenizer(br.readLine(), " ");
         
-        int a = Integer.parseInt(st.nextToken());
-        int b = Integer.parseInt(st.nextToken());
+        long a = Integer.parseInt(st.nextToken());
+        long b = Integer.parseInt(st.nextToken());
         
         a = a*a;
         b = b*b;
